+++ HeExample/HeControl/HAbstractClientWidget.h
@@ -16,7 +16,7 @@ class HAbstractClientWidget : public HAbstractNetworkWidget
     Q_OBJECT
     Q_DECLARE_PRIVATE(HAbstractClientWidget)
     Q_PROPERTY(QString serverIP READ serverIP WRITE setServerIP)
-    Q_PROPERTY(quint16 serverPort READ serverPort WRITE setServerPort)
+    Q_PROPERTY(int serverPort READ serverPort WRITE setServerPort)
 
 public:
     explicit HAbstractClientWidget(QWidget *parent = nullptr);
@@ -24,11 +24,11 @@ public:
 
 public:
     QString serverIP() const;
-    quint16 serverPort() const;
+    int serverPort() const;
 
 public slots:
     void setServerIP(const QString &value);
-    void setServerPort(quint16 value);
+    void setServerPort(int value);
 
 protected:
     HAbstractClientWidget(HAbstractClientWidgetPrivate &p, QWidget *parent = nullptr);
