+++ HeExample/HeControl/HUdpServer.h
@@ -21,10 +21,10 @@ public:
     ~HUdpServer() override;
 
 signals:
-    void clientConnected(const QString &ip, quint16 port);
-    void clientDisconnected(const QString &ip, quint16 port);
-    void sentData(const QString &ip, quint16 port, const QByteArray &data);
-    void receiveData(const QString &ip, quint16 port, const QByteArray &data);
+    void clientConnected(const QString &ip, int port);
+    void clientDisconnected(const QString &ip, int port);
+    void sentData(const QString &ip, int port, const QByteArray &data);
+    void receiveData(const QString &ip, int port, const QByteArray &data);
 
 public slots:
     // 是否运行
@@ -36,15 +36,15 @@ public slots:
     // 设置监听IP
     void setListenIP(const QString &value);
     // 设置监听端口
-    void setListenPort(quint16 value);
+    void setListenPort(int value);
     // 发送数据
     void sendData(const QByteArray &data);
-    void sendData(const QString &ip, quint16 port, const QByteArray &data);
+    void sendData(const QString &ip, int port, const QByteArray &data);
     // 新连接
-    void incomingConnection(const QString &ip, quint16 port);
+    void incomingConnection(const QString &ip, int port);
     // 断开连接
     void disconnectClient();
-    void disconnectClient(const QString &ip, quint16 port);
+    void disconnectClient(const QString &ip, int port);
 
 protected:
     HUdpServer(HUdpServerPrivate &p, QObject *parent = nullptr);
