+++ HeExample/HeControl/HUdpServerWidget.h
@@ -24,15 +24,15 @@ public:
     explicit HUdpServerWidget(QWidget *parent = nullptr);
     ~HUdpServerWidget() override;
 
-public slots:    
+public slots:
     void sendData() override;
     void clearData() override;
 
 protected slots:
     void on_pushButton_101_clicked();
     void on_pushButton_102_clicked();
-    void handleClientConnected(const QString &ip, quint16 port) override;
-    void handleClientDisconnected(const QString &ip, quint16 port) override;
+    void handleClientConnected(const QString &ip, int port) override;
+    void handleClientDisconnected(const QString &ip, int port) override;
     void append(int type, QString data) override;
 
 protected:
