+++ HeExample/HeControl/HUdpServer_p.h
@@ -13,7 +13,7 @@ class HUdpServerPrivate
 public:
     bool isRunning = false;
     QString listenIP = "0.0.0.0";
-    quint16 listenPort = 6000;
+    int listenPort = 6000;
     QSet<QString> clients;
     QUdpSocket *socket;
 };
